Renamed error_utils.h to error_utils.hpp since it's a CPP header.
