+++ cpp/src/binary/binary_ops.cu
@@ -2,7 +2,7 @@
 
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "utilities/nvtx/nvtx_utils.h"
