+++ cpp/src/bitmask/bitmask_ops.cu
@@ -1,6 +1,6 @@
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "cudf/functions.h"
 #include "rmm/thrust_rmm_allocator.h"
