+++ cpp/src/compaction/stream_compaction_ops.cu
@@ -31,7 +31,7 @@
 
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "rmm/thrust_rmm_allocator.h"
 
 //std lib
