+++ cpp/src/join/joining.cu
@@ -21,7 +21,7 @@
 
 #include "cudf.h"
 #include "rmm/rmm.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "dataframe/cudf_table.cuh"
 #include "utilities/nvtx/nvtx_utils.h"
