+++ cpp/src/orderby/orderby.cu
@@ -20,7 +20,7 @@
 
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 
 #include "rmm/thrust_rmm_allocator.h"
