+++ cpp/src/quantiles/quantiles.cu
@@ -21,7 +21,7 @@
 
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "rmm/thrust_rmm_allocator.h"
 
 #include "quantiles.h"
