+++ cpp/src/reductions/reductions.cu
@@ -1,6 +1,6 @@
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "utilities/type_dispatcher.hpp"
 
 #include <cub/block/block_reduce.cuh>
