+++ cpp/src/reductions/scan.cu
@@ -1,7 +1,7 @@
 #include "cudf.h"
 #include "rmm/rmm.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "bitmask/bit_mask.cuh"
 #include "utilities/type_dispatcher.hpp"
