+++ cpp/src/sort/segmented_sorting.cu
@@ -1,7 +1,7 @@
 #include "cudf.h"
 #include "rmm/rmm.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 
 
 #include <cub/device/device_segmented_radix_sort.cuh>
