+++ cpp/src/sqls/sqls_ops.cu
@@ -4,7 +4,7 @@
 
 #include "cudf.h"
 #include "utilities/cudf_utils.h"
-#include "utilities/error_utils.h"
+#include "utilities/error_utils.hpp"
 #include "rmm/thrust_rmm_allocator.h"
 
 #include "sqls_rtti_comp.h"
