+++ cpp/src/utilities/error_utils.hpp
